+++ src/stats/fault.jl
@@ -7,7 +7,7 @@ Gets fault currents for switches and corresponding fault from study in-place in 
 using [`get_timestep_fault_currents`](@ref get_timestep_fault_currents).
 """
 function get_timestep_fault_currents!(args::Dict{String,<:Any})::Vector{Dict{String,Any}}
-    args["output_data"]["Fault currents"] = get_timestep_fault_currents(get(args, "fault_studies_results", Dict{String,Any}()), get(args, "faults", Dict{String,Any}()), args["network"])
+    args["output_data"]["Fault currents"] = get_timestep_fault_currents(get(args, "fault_studies_results", Dict{String,Any}()), get(args, "faults", Dict{String,Any}()), args["network"]; ret_protection_only=!(get(args, "debug", false) || get(args, "verbose", false)))
 end
 
 
@@ -15,7 +15,8 @@ end
     get_timestep_fault_currents(
         fault_studies_results::Dict{String,<:Any},
         faults::Dict{String,<:Any},
-        network::Dict{String,<:Any}
+        network::Dict{String,<:Any};
+        ret_protection_only::Bool=false
     )::Vector{Dict{String,Any}}
 
 Gets information about the results of fault studies at each timestep, including:
@@ -31,8 +32,11 @@ Gets information about the results of fault studies at each timestep, including:
   - the positive-sequence fault current `|I1| (A)`
   - the negative-sequence fault current `|I2| (A)`
   - the bus voltage from the from-side of the switch `|V| (V)`
+
+`ret_protection_only==false` indicates that currents and voltages should be returned for all lines where switch=y, and if
+`true`, should only return switches for which a protection device is defined (recloser, relay, fuse)
 """
-function get_timestep_fault_currents(fault_studies_results::Dict{String,<:Any}, faults::Dict{String,<:Any}, network::Dict{String,<:Any})::Vector{Dict{String,Any}}
+function get_timestep_fault_currents(fault_studies_results::Dict{String,<:Any}, faults::Dict{String,<:Any}, network::Dict{String,<:Any}; ret_protection_only::Bool=true)::Vector{Dict{String,Any}}
     fault_currents = Dict{String,Any}[]
 
     for n in sort([parse(Int, i) for i in keys(fault_studies_results)])
@@ -67,6 +71,17 @@ function get_timestep_fault_currents(fault_studies_results::Dict{String,<:Any}, 
                             ) for (id, switch) in get(network["nw"]["$n"], "switch", Dict())
                         ),
                     )
+
+                    if ret_protection_only
+                        protection_locations = Set()
+                        for pt in _pnm2eng_objects["protection"]
+                            for (i,pd) in get(network["nw"]["$n"], pt, Dict())
+                                haskey(pd, "location") && !isempty(pd["location"]) && push!(protection_locations, string(replace(pd["location"], "line."=>"")))
+                            end
+                        end
+
+                        _fault_currents[bus_id][fault_type][fault_id]["switch"] = filter(x->x.first∈protection_locations, _fault_currents[bus_id][fault_type][fault_id]["switch"])
+                    end
                 end
             end
         end
@@ -85,7 +100,7 @@ end
 
 Special case where the faults string was not parsed
 """
-get_timestep_fault_currents(fault_studies_results::Dict{String,<:Any}, faults::String, network::Dict{String,<:Any})::Vector{Dict{String,Any}} = get_timestep_fault_currents(fault_studies_results, Dict{String,Any}(), network)
+get_timestep_fault_currents(fault_studies_results::Dict{String,<:Any}, faults::String, network::Dict{String,<:Any}; ret_protection_only::Bool=false)::Vector{Dict{String,Any}} = get_timestep_fault_currents(fault_studies_results, Dict{String,Any}(), network; ret_protection_only=ret_protection_only)